IgScan: A high-throughput analysis toolkit for dissecting immune repertoires in B-cell neoplasms and beyond

IgScan is an R package for processing and analyzing high-throughput sequencing data of B-cell immunoglobulin (Ig) repertoires, both at the bulk and single-cell levels. The single-cell workflow supports multiple data formats, including 10x Genomics, BD Rhapsody, and Parse Evercode. IgScan provides detailed immunogenetic annotation, clonotype grouping, and repertoire reconstruction, with open compatibility with AIRR standards and other widely used tools in the field.

Useful links

Package website: https://nadeulab.github.io/IgScan/

Report bugs at: https://github.com/nadeulab/IgScan/issues

See also

Author

Maintainer: Ian Márquez-López iamarquez@recerca.clinic.cat

Authors:

Ian Márquez-López

Ferran Nadeu nadeu@recerca.clinic.cat