Run IgScan Annotation

This function performs the IgScan annotation on Run_IgBlast_from_RawData output data. It groups sequences into clonotypic clusters based on a highly customizable hierarchical clustering approach. The function processes clonotypes to provide key immunogenetic information, such as germline sequences, somatic hypermutation levels, CLL stereotype subsets, R110 mutation, Acquired N-Glycosylation sites (AGS) and more. Then, it organizes this information according to the input data type: for single-cell datasets, it groups heavy-light chains by cell barcodes, while for bulk-NGS datasets, it reports the cumulative number of reads and clonotype frequencies.

Run_IgScan_Annotation(
  sample_labels = "all_samples",
  case_labels = NULL,
  input_format,
  outputDir,
  analysis_mode = "single",
  material_type = "rna",
  v_primer = "full_length",
  data_type = "single_cell",
  min_reads = 2,
  remove_tmp = TRUE,
  hc_similarity_cutoff = 0.2,
  hc_mode = "average",
  cdr3_mode = "nt",
  cdr3_InDel_correction_mode = "soft_filter",
  annotate_CLL_immGen = FALSE,
  annotate_satellite_subsets = TRUE,
  annotate_ags = FALSE,
  summary_file = NULL,
  threads = 1
)

Arguments

sample_labels

A vector of sample labels to be annotated. Must match the sample labels of the Run_IgBlast_from_RawData output files. This parameter can be also set to 'all_samples', which will analyze all the samples in the igblast_out directory in 'single' analysis_mode. Default is 'all_samples'.

case_labels

A vector of case labels matching the indexes of sample_labels. Required when analysis mode is set to 'joint'. Default is NULL.

input_format

A string specifying the format of the input data, currently supporting: '10xBCR_fasta', '10xBCR_csv', 'ParseBCR', 'BDRhapsodyBCR', 'MiXCR', 'TRUST4', 'AIRR', 'IMGT_AIRR' and 'fasta'.

outputDir

Path to the directory containing the previous IgScan outputs coming from the Run_IgBlast_from_RawData function. There is NO default value for this parameter.

analysis_mode

Defines the mode of analysis to be performed. Options are "single" and "joint". In "single" mode, each sample is annotated independently. In "joint" mode, samples are grouped by case based on the "case_list" vector for a case-level joint annotation. Default is 'single'.

material_type

The biological source of material. Options: 'dna' and 'rna'. Note that for matrial_type='rna', unproductive sequences are not expected, and will be directly removed. Default is 'rna'.

v_primer

The primer sequence used for the V-region amplification. Options: 'full_length', 'fr1', 'fr2' and 'fr3'. Note that sequences with unexpected length pattern based on the chosen primer will be directly excluded from the analysis. Default is 'full_length'.

data_type

The type of data. Options: 'single_cell' and 'bulk'. Default is 'single_cell'.

min_reads

The minimum number of reads required to keep a sequence in the downstream analyses. Only needed for bulk NGS data. Default is 2.

remove_tmp

Logical value indicating whether to remove the temporary files (all files except the annotation_results directory). Default is TRUE.

hc_similarity_cutoff

Hierarchical clustering distance in CDR3 identity to consider two sequences clonally related. Default is 0.2, meaning that up to 20 percent of mismatch is accepted.

hc_mode

The aglomeration method to be used in the hierarchical clustering. Options are: 'single', 'complete' and 'average' (UPGMA). Default is 'average'.

cdr3_mode

The type of CDR3 sequence to be used for the hierarchical clustering. Options are 'nt' (nucleotides) and 'aa' (amino acids). Default is 'nt'.

cdr3_InDel_correction_mode

The degree of stringency in correcting InDels in the CDR3 region. Options are 'no' (no CDR3 InDel correction, recommended for highly polyclonal samples), 'hard_filter' (high stringency but computationally slower) and 'soft_filter' (less stringent but much faster). Default is 'soft_filter'.

annotate_CLL_immGen

Logical value indicating whether to annotate CLL-related immunogenetic information such as CLL stereotyped subsets and the R110 mutation. Default is FALSE.

annotate_satellite_subsets

Logical value indicating whether to annotate Satellite CLL stereotyped subsets. Only needed if annotate_CLL_immGen is set to TRUE. Default is TRUE.

annotate_ags

Logical value indicating whether to annotate IGH Acquired N-Glycosylation Sites (AGS). Default is FALSE.

threads

An integer specifying the number of threads for processing samples. Default is 1.

Value

A list containing an IgScan annotation dataframe for every sample analyzed, which are also saved in the output directory. These dataframes can be directly passed to further IgScan functions.

Examples

if (FALSE) { # \dontrun{
# Example 1: XXXXX
result <- Run_IgScan_Annotation(analysis_mode = "joint",
   sample_labels = c("Sample1", "Sample2"),
   case_labels = c("Case1", "Case1"),
   input_format = "fasta",
   material_type = "dna",
   v_primer = "full_length",
   data_type = "bulk",
   min_reads = 2,
   remove_tmp = TRUE,
   outputDir = "path/output/dir/",
   hc_similarity_cutoff = 0.2,
   hc_mode = "average",
   cdr3_mode = "nt",
   cdr3_InDel_correction_mode = "soft_filter",
   annotate_CLL_immGen = TRUE,
   annotate_satellite_subsets = FALSE,
   annotate_ags = TRUE,
   threads = 4)

# Example 2: XXXXX
result <- Run_IgScan_Annotation(
   analysis_mode = "joint",
   sample_labels = c("Sample1", "Sample2"),
   case_labels = c("Case1", "Case1"),
   input_format = "fasta",
   material_type = "dna",
   v_primer = "full_length",
   data_type = "bulk",
   min_reads = 2,
   remove_tmp = TRUE,
   outputDir = "path/output/dir/",
   hc_similarity_cutoff = 0.2,
   hc_mode = "average",
   cdr3_mode = "nt",
   cdr3_InDel_correction_mode = "soft_filter",
   annotate_CLL_immGen = TRUE,
   annotate_satellite_subsets = FALSE,
   annotate_ags = TRUE,
   threads = 4)
} # }