Combine IgScan Output with a Single Cell Experiment Object

This function integrates immunogenetic data from IgScan with a Single Cell Experiment (SCE) object. It maps clonotype-related information and sequence data to the colData of the SCE object based on matching barcodes.

combine_IgScan_SingleCellExperiment(
  igscan_out,
  sce,
  sce_sample_col,
  igscan_sample_col = "SampleID",
  threads = 1
)

Arguments

igscan_out

A data frame containing the output from IgScan for a single cell dataset.

sce

A SCE object where colData will be updated with the corresponding immunogenetic information from igscan_out.

sce_sample_col

A vector with the name of the column or columns containing the sample identifiers in the metadata of the SCE object. There is no default value for this parameter.

igscan_sample_col

A vector with the name of the column or columns containing the sample identifiers in the IgScan annotation data frame. Default is 'SampleID'.

threads

The number of threads to be used. Default is 1.

Value

A SCE object with updated colData, including the IgScan annotation organized by cell barcode.

Examples

if (FALSE) { # \dontrun{
updated_sce_object <- combine_IgScan_Seurat(igscan_out, sce_object, sce_sample_col = "Sample", igscan_sample_col = "SampleID", threads = 4)
} # }