Export IgScan outputs in AIRR format

This function converts the output from IgScan (or compatible objects single cell objects with IgScan annotation) into the standardized AIRR format and writes it to a .tsv file. It supports input objects from IgScan output data frame, Seurat, or SingleCellExperiment classes.

export_AIRR_format(
  object,
  dir = NULL,
  fileName = NULL,
  germline_aln = "masked",
  sample_col = "SampleID",
  metadata = NULL
)

Arguments

object

A data frame, Seurat object, or SingleCellExperiment object containing IgScan annotation fields. For data frames, it is expected to match the IgScan format (specific column names will be validated).

dir

A character string specifying the output directory. If it does not exist, it will be created. Default is NULL, meaning that the AIRR file will not be written to a file, it will only be returned by the function as a data.frame.

fileName

A character string for the output file name. Defaults to "IgScan_AIRR_formatted.tsv" if not provided.

germline_aln

Type of germline alignment to be included in the AIRR file. Options are masked (uses IgBlast masked germline alignment) and consensus (uses the IgScan custom-built consensus germline sequence).

sample_col

A vector with the name of the column containing the sample identifier. Default is 'SampleID'.

metadata

A character vector specifying the columns in the object that should also be included in the AIRR file as metadata. Default is NULL.

Value

A .tsv file written to the specified directory in AIRR format as well as the AIRR dataframe as an R object.

Examples

if (FALSE) { # \dontrun{
# Export from a data.frame
export_AIRR_format(igscan_df, dir = "results/")

# Export from Seurat object
export_AIRR_format(seurat_obj, dir = "results/", germline_aln = "consensus")
} # }