Prepare gctree inputs from IgScan annotated datasets

This function takes as input an IgScan annotated dataframe coming from bulk NGS or single cell workflows, or a single cell object with IgScan annotation, and prepares the input files for running the gctree program for BCR phylogenetic inference. Once the input files have been generated, the user can run gctree following the instructions in XXXX jupyter notebook provided with the IgScan package.

prepare_gctree_input(
  object,
  outDir,
  mode = "single_cell",
  cloneID = "C1",
  chain = "HL"
)

Arguments

object

A data.frame, SingleCellExperiment, or Seurat object with IgScan annotation.

outDir

A string. Path to the directory where output files should be written.

mode

A string. Either "single_cell" or "bulk" depending on the dataset type. Default is "single_cell".

cloneID

A string. The clonotype identifier to extract and process. Note that the cloneID parameter must match a value in the clonotype identifier field in the IgScan objects.

chain

A string. One of "H" (heavy chain), "L" (light chain), or "HL" (paired chains). Default is "HL". Only needed when mode is set to "single_cell".

Value

The function returns no value in R but writes the necessary input files for gctree into the outDir.

Details

The function generates the following files in the specified outDir:

abundances.csv: Abundance of each subclone including the germline (GL).

idmap.txt: Mapping of subclone IDs to individual barcodes (or rownames).

subclones.fasta: FASTA file of the aligned somatic and germline VDJ sequences.

deduplicated.phylip: Alignment file in PHYLIP format for input to gctree.

Examples

if (FALSE) { # \dontrun{
# Example with a SingleCellExperiment object
prepare_gctree_input(sce_object, outDir = "results/clone1/", mode = "single_cell", cloneID = "C1", chain = "HL")

# Example with a Seurat object
prepare_gctree_input(seurat_object, outDir = "results/clone2/", mode = "single_cell", cloneID = "C5", chain = "H")

# Example with a bulk IgScan-annotated dataframe
prepare_gctree_input(bulk_df, outDir = "results/clone3/", mode = "bulk", cloneID = "CloneA")
} # }